CRAYSAT-1133: Add documentation
This commit adds Sphinx documentation that uses the autodoc extension
to automatically document the ProductCatalog class. It also uses
the napoleon extension to understand Google-stype doc strings.

Also, add the Sphinx build directory to .gitignore.

Finally, a README.md is added with a reference to the Sphinx docs,
a reference to the design doc, and some basic instructions to get
started.