CRAYSAT-1161: Update pip URL to point at arti
This change is necessary since artifactory.algol60.net does not resolve
within CJE. This is an issue with DNS within GCP and may get resolved in
the future. For the time being, there is an internal mirror on arti
which hosts points to the cray-product-catalog packages.

Test Description:
Run `make pymod_test`.