CRAYSAT-1371: Add HTTP authentication support
In CSM 1.2, access to Nexus requires HTTP authentication. This is
supported by Nexusctl, the library we are using to interact with Nexus.
However, Nexusctl requires the credentials to be set in the environment,
and they can be found in a Kubernetes secret.

This commit adds code to the constructor method of ProductCatalog to
read the credentials stored in the Kubernetes secret and updates the
environment variables as is expected by Nexusctl.

Additionally, the `parser` module is enhanced to provide CLI arguments
that may override the name and namespace of this secret, should it
change in the future.

Finally this commit adds unit tests and a changelog entry. This commit
defines version 2.3.0.

Test Description:
* I pushed this branch, resulting in the Jenkins pipeline publishing
  a dev version to internal-pip-unstable-local.
* Next, I built a version of `sat-install-utility` that included this
  version of `shasta-install-utility-common` (and in turn the new
  version of `nexusctl`).
* I used `prodmgr` to uninstall and activate mock versions of SAT on mug.