CRAYSAT-1750: Relax cray-product-catalog constraint
Currently, `sat-install-utility` depends on cray-product-catalog >=
1.6.0 and on `shasta-install-utility-common`. However, since
`shasta-install-utility-common` depends on exactly ersion 1.3.2 of
`cray-product-catalog`, these dependencies can't be resolved. Fix this
by allowing the `cray-product-catalog` version required by this package
to be >= 1.3.2 instead of == 1.3.2.

Consistent with semver, the changes between 1.3.2 and 1.8.12 of
cray-product-catalog seem to just be additions. The interfaces haven't
change in a backwards-incompatible way.

One change is needed to a unit test. `cray-product-catalog` relaxed the
schema of product catalog data, so that products can provide additional
properties. A unit test that checked for this being disallowed is
removed here.